From Martin Kraemer:
Such an "ancient" locale used (in most countries where applicable,
e.g. germany, scandinavian countries etc) the ASCII-characters
'{', '|', '}', '[', '\\', ']', and '~' for displaying local
umlaut characters like adiaresis, odiaresis, udiaresis or "sharp s",
and the like. So, these characters have the bad property of belonging
both to tcsh's punctuation characters, and to the alphanumeric class...
csh and tcsh have in the past very much disliked such a locale, and
the effect was for instance:

> echo ${shell}
Missing '}'

Reason is that in such a locale, the ctype function return true
for isalpha('{'), so much of tcsh's logic breaks.

Don't return true for isalpha and isalnum if it is a punctuation character.