1. $promptchars
What does it mean anyway when the user sets $promptchars to an empty
string (or just one char)?  Do they want their prompt to be empty?
If so, they would certainly fare better if they set their $prompt (not
$promptchars) to be emtpy, wouldn't they?  Did they just do it by
mistake?  Then, they could use some warning of sorts.

Another thing is that ordinary users probably wouldn't want to bother
about what kind of prompt they'd get as a root, but the $promptchars
shell variable sort of requires them to, or at least it used to.


If PRCH (or PRCHROOT) is '\0', the current handling of $prompt makes
everything after '%#' disappear, or if attributes are involved, makes
'%#' appear as '^@' (because a null with attributes != '\0'):

	> set prompt='(%#) '
	(>) set promptchars
	(set prompt='%B(%#)%b '
	(^@) _

This doesn't look very pretty, does it?  It's certainly not informative
enough to give a novice user a clue as to what has happened here.

My initial patch was careful not to assign a null character to PRCH or
PRCHROOT specifically to avoid this problem, setting them to their
default values if $promptchars is empty or contains only one character.

But there is another solution to this.  We could change the tprintf()
routine in tc.prompt.c so that it just ignores the %# format string if
PRCH (or PRCHROOT) is '\0'.  This would perhaps give the user a better
visual clue as to what went wrong.  Here's a patch for this: