autoset kanji changes The patch attached provides a solution.  I have reused the shell variable nokanji, to have a boolean around to be able to skip the translation of the Shift-JIS characters Yen (which is backslash aka 0x5c in ASCII) and an overline (which is in fact the place at 0x7e of ASCII tilde).  Also I make it possible to use Meta if MB_CUR_MAX is 1 which is the case of using POSIX only.
For not to stumble on two byte characters using byte 0x5c or
byte 0x7e the last returned state of mbrtowc(3) is required.
Therefore the mbstate_t variable has to be a static one.
Maybe the memset to zero below could be skipped as this may
provide a better support of multi byte characters becoming
disconnected over buffer boundaries.