Remove windows and exclude gcc on osx
Apparently calling nmake directly is not the way to go, but instead
we should be calling msbuild. I don't know how to figure that out.

https://travis-ci.community/t/nmake-support-from-windows/1079

The gcc command on macOS (osx) is just another name for clang, so
let's not run the same compile twice unnecessarily. It is probably
not that interesting to try to install gcc and run that on macOS,
so I'm not doing it.

https://docs.travis-ci.com/user/languages/c/#gcc-on-macos