implement CLICOLOR_FORCE
Support environment variable CLICOLOR_FORCE to output
colors in ls-F even if the output isn't a terminal.

The behaviour is based on the FreeBSD/macOS ls(1) which uses
the presence of $CLICOLOR_FORCE and does not parse the value.

This will allow the testsuite to be extended to test
various color features.

Minor style fixes, including changing the existing
color_as_referent to be static since it's not used
outside of tw.color.c.