Use $+ instead of $^ in nls/Makefile.in
NetBSD make added support for $^ per POSIX, so it no longer expands to
an empty string, which breaks the rules.  It does not define $+, so we
can use that instead for GNU make.

A proper fix would be to detect the make flavor used, and build rules
appropriate for each flavor.  Listing all the dependencies is not
practical (in the rules involving multiple set files).